Extract ISO_8859_15 to its own file
